% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-analyte.R
\name{Analyte}
\alias{Analyte}
\title{Analyte Object}
\description{
Analyte class is an object that gathers information about one specific
Analyte on the Luminex plate
}
\examples{
etramp <- Analyte$new(id = 73, analyte_name = "Etramp5_ag1", bead_count = 50)
print(etramp)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{A numeric value that represents the unique identifier
of the analyte}

\item{\code{analyte_name}}{A character value that represents the name of the analyte}

\item{\code{bead_count}}{A numeric value that represents the number of beads}

\item{\code{analysis_type}}{Type of the analyte}

\item{\code{units}}{Units of the analyte in which the results are expressed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Analyte-new}{\code{Analyte$new()}}
\item \href{#method-Analyte-print}{\code{Analyte$print()}}
\item \href{#method-Analyte-join}{\code{Analyte$join()}}
\item \href{#method-Analyte-clone}{\code{Analyte$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte-new"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{Analyte} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte$new(
  id,
  analyte_name,
  bead_count = NA,
  analysis_type = NULL,
  units = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{numeric(1)})\cr
A numeric value that represents the unique identifier of the analyte}

\item{\code{analyte_name}}{(\code{character(1)})\cr
A character value that represents the name of the analyte}

\item{\code{bead_count}}{(\code{numeric(1)})\cr}

\item{\code{analysis_type}}{(\code{character(1)})\cr
Type of the analyte}

\item{\code{units}}{(\code{character(1)})\cr
Units of the analyte in which the results are expressed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte-print"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte-print}{}}}
\subsection{Method \code{print()}}{
Prints the information about the analyte
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte-join"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte-join}{}}}
\subsection{Method \code{join()}}{
Joins the data of two analytes
Performs checks to ensure that the analytes can be joined
If the analytes can be joined, the bead_count, analysis_type and units
are updated in the current analyte object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte$join(new_analyte)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_analyte}}{(\code{Analyte})\cr
The analyte object to be joined with the current analyte object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated analyte object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
