% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_path_equal}
\alias{check_path_equal}
\title{Check if two paths are equal}
\usage{
check_path_equal(path1, path2)
}
\arguments{
\item{path1}{(\code{character(1)}) The first path to be compared.}

\item{path2}{(\code{character(1)}) The second path to be compared.}
}
\value{
(\code{logical(1)}) \code{TRUE} if the paths are equal, \code{FALSE} otherwise.
}
\description{
Function checks if two paths are equal after converting them to absolute paths.
}
\keyword{internal}
