% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-plate_builder.R
\name{convert_dilutions_to_numeric}
\alias{convert_dilutions_to_numeric}
\title{Convert dilutions to numeric values}
\usage{
convert_dilutions_to_numeric(dilutions)
}
\arguments{
\item{dilutions}{vector of dilutions used during the examination saved
as strings in format \verb{1/\\d+}}
}
\value{
a vector of numeric values representing the dilutions
}
\description{
Convert dilutions saved as strings in format \verb{1/\\d+} into numeric values
}
\keyword{internal}
