% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{create_standard_curve_model_analyte}
\alias{create_standard_curve_model_analyte}
\title{Create standard curve model for a certain analyte}
\usage{
create_standard_curve_model_analyte(
  plate,
  analyte_name,
  data_type = "Median",
  ...
)
}
\arguments{
\item{plate}{Plate object}

\item{analyte_name}{Name of the analyte for which we want to create the model}

\item{data_type}{Data type of the value we want to use to fit the model - the same datatype as in the plate file. By default equals to \code{Median}}

\item{...}{Additional arguments passed to the model}
}
\value{
Standard Curve model
}
\description{
Create standard curve model for a certain analyte
}
