% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-plate.R
\name{create_vector_without_holes}
\alias{create_vector_without_holes}
\title{Remove holes from a vector}
\usage{
create_vector_without_holes(vector, locations)
}
\arguments{
\item{vector}{A vector with values}

\item{locations}{A vector with locations where the values should be placed}
}
\value{
A vector with values at the given locations
}
\description{
Function selects the values from the vector at the given locations
and creates a vector "without holes". Works only for 96-plate.
}
\keyword{internal}
