% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-dir.R
\name{detect_mbr_format}
\alias{detect_mbr_format}
\title{Try to detect the format of a file}
\usage{
detect_mbr_format(filepath, format = NULL)
}
\description{
Try to detect the format of a file
}
\keyword{internal}
