% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dilution_to_rau}
\alias{dilution_to_rau}
\title{Convert dilution to RAU}
\usage{
dilution_to_rau(predicted_dilution)
}
\arguments{
\item{predicted_dilution}{(\code{numeric()}) A numeric value representing the predicted dilution.}
}
\value{
The RAU value corresponding to the predicted dilution .
}
\description{
Convert dilution to RAU
}
\keyword{internal}
