% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-plate_builder.R
\name{extract_dilution_from_names}
\alias{extract_dilution_from_names}
\title{Extract dilution factor from the sample name}
\usage{
extract_dilution_from_names(sample_name)
}
\arguments{
\item{sample_name}{a vector of sample names from which we want to extract the dilutions}
}
\value{
a vector of dilutions represented as strings extracted from the sample names
}
\description{
function extracts dilution factor from the sample name - useful for detecting
dilution from sample names
}
\keyword{internal}
