% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-plate_builder.R
\name{extract_sample_names_from_layout}
\alias{extract_sample_names_from_layout}
\title{Extract sample names from layout}
\usage{
extract_sample_names_from_layout(layout_names, locations)
}
\description{
Function extracts sample names from the layout file based on the provided locations.
Function assumes that the plate is 96-well and extracts
the sample names according to the provided location strings.
}
\examples{
layout_names <- paste0(c("SAMPLE"), 1:96)
locations <- c("A1", "A2", "A3", "B4")
extract_sample_names_from_layout(layout_names, locations)
}
