% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-dir.R
\name{find_layout_file}
\alias{find_layout_file}
\title{Find a layout file given plate filepath}
\usage{
find_layout_file(plate_filepath, layout_filepath = NULL)
}
\description{
Find a layout file given plate filepath
}
\keyword{internal}
