% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_plate_report}
\alias{generate_plate_report}
\title{Generate a report for a plate.}
\usage{
generate_plate_report(
  plate,
  use_model = TRUE,
  filename = NULL,
  output_dir = "reports",
  counts_lower_threshold = 50,
  counts_higher_threshold = 70,
  additional_notes = NULL
)
}
\arguments{
\item{plate}{A plate object.}

\item{use_model}{(\code{logical(1)}) A logical value indicating whether the model should be used in the report.}

\item{filename}{(\code{character(1)}) The name of the output file. If not provided,the filename will be created based on the plate name with the suffix '_report.html'.}

\item{output_dir}{(\code{character(1)}) The directory where the report should be saved. Default is 'reports'.}

\item{counts_lower_threshold}{(\code{numeric(1)}) The lower threshold for the counts plots (works for each analyte). Default is 50.}

\item{counts_higher_threshold}{(\code{numeric(1)}) The higher threshold for the counts plots (works for each analyte). Default is 70.}

\item{additional_notes}{(\code{character(1)}) Additional notes to be included in the report. Contents of this fields are left to the user's discretion. If not provided, the field will not be included in the report.}
}
\value{
A report.
}
\description{
This function generates a report for a plate. The report contains all the necessary information about the plate, from the general plate parameters, such as examination date, to the breakdown of the analytes' plots.
The report is generated using the \code{plate_report_template.Rmd} template.
}
\examples{

plate_file <- system.file("extdata", "CovidOISExPONTENT.csv", package = "PvSTATEM")
layout_file <- system.file("extdata", "CovidOISExPONTENT_layout.csv", package = "PvSTATEM")
note <- "This is a test report.\n**Author**: Jane Doe \n**Tester**: John Doe"

plate <- read_luminex_data(plate_file, layout_file)
tmp_dir <- tempdir(check = TRUE)
generate_plate_report(plate,
  output_dir = tmp_dir,
  counts_lower_threshold = 40,
  counts_higher_threshold = 50,
  additional_notes = note
)
}
