% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate a report for a plate.}
\usage{
generate_report(plate, use_model = TRUE)
}
\arguments{
\item{plate}{A plate object.}

\item{use_model}{A logical value indicating whether the model should be used in the report.}
}
\value{
A report.
}
\description{
This function generates a report for a plate. The report is generated using
the \code{html_report_template.Rmd} template.
}
