% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_join_value}
\alias{get_join_value}
\title{Determine the Join Value}
\usage{
get_join_value(x, y)
}
\arguments{
\item{x}{A value to be compared.}

\item{y}{A value to be compared.}
}
\value{
A non-\code{NA}/non-\code{NULL} value or the common value if \code{x} equals \code{y}. Returns \code{NULL} if the values differ and neither is \code{NA} or \code{NULL}.
}
\description{
Returns a non-\code{NA}/non-\code{NULL} value based on the inputs. If either value is \code{NA} or \code{NULL}, it returns the non-\code{NA}/non-\code{NULL} value. If both values are equal, it returns that value.
}
\keyword{internal}
