% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-plate_builder.R
\name{get_location_matrix}
\alias{get_location_matrix}
\title{Generate the matrix of plate locations}
\usage{
get_location_matrix(nrow = 8, ncol = 12, as_vector = FALSE)
}
\arguments{
\item{nrow}{Number of rows in the plate}

\item{ncol}{Number of columns in the plate}

\item{as_vector}{logical value indicating whether to return the locations as a vector}
}
\value{
a matrix with locations
}
\description{
The function generates a matrix of plate locations. The locations are represented in a nrow x ncol matrix.
Usually number of rows equals to 8 and number of columns to 12, and the total matrix size is 96.

The fields are represented as \code{E3}, where the letter corresponds to the row and the number to the column.
}
\keyword{internal}
