% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-dir.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Get output directory for a given input file}
\usage{
get_output_dir(
  input_file,
  input_dir,
  output_dir = NULL,
  flatten_output_dir = FALSE
)
}
\description{
Get output directory for a given input file
}
\keyword{internal}
