% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{handle_datetime}
\alias{handle_datetime}
\title{Handle differences in datetimes}
\usage{
handle_datetime(datetime_str, file_format = "xPONENT")
}
\arguments{
\item{datetime_str}{The datetime string to parse}

\item{file_format}{The format of the file. Select from: xPONENT, INTELLIFLEX}
}
\value{
POSIXct datetime object
}
\description{
Handle differences in the datetime format between xPONENT and INTELLIFLEX
and output POSIXct datetime object containing the correct datetime with the default timezone.
}
\keyword{internal}
