% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.decreasing}
\alias{is.decreasing}
\title{Check if the vector is monotically decreasing}
\usage{
is.decreasing(x)
}
\arguments{
\item{x}{(\code{numeric()}) Vector of numeric values}
}
\value{
(\code{logical(1)}) \code{TRUE} if the vector is monotonically decreasing, \code{FALSE} otherwise
}
\description{
Check if the vector is monotically decreasing
}
\keyword{internal}
