% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.scalar}
\alias{is.scalar}
\title{Check if a value is a scalar}
\usage{
is.scalar(x)
}
\arguments{
\item{x}{Object to be checked.}
}
\value{
\code{TRUE} if the object is a scalar, \code{FALSE} otherwise.
}
\description{
This will return FALSE for NULL and vectors of length bigger than 2.
}
\keyword{internal}
