% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.str.number}
\alias{is.str.number}
\title{Check if a string is a number}
\usage{
is.str.number(x)
}
\arguments{
\item{x}{A string to be checked.}
}
\value{
\code{TRUE} if the string is a number, \code{FALSE} otherwise.
}
\description{
Check if a string is a number
}
\keyword{internal}
