% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-dir.R
\name{is_mba_data_file}
\alias{is_mba_data_file}
\title{Identify if a file is a MBA data file}
\usage{
is_mba_data_file(filepath, check_format = TRUE)
}
\arguments{
\item{filepath}{(\code{character(1)}) The path to the file.}

\item{check_format}{(\code{logical(1)}) If \code{TRUE}, the function will check if the file name contains a supported format. The default is \code{TRUE}.}
}
\value{
\code{TRUE} if the file is a MBA data file, \code{FALSE} otherwise.
}
\description{
Identify if a file is a MBA data file
}
\keyword{internal}
