% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_mba_format}
\alias{is_mba_format}
\title{Check if a mba format is supported}
\usage{
is_mba_format(format, allow_nullable = FALSE)
}
\arguments{
\item{format}{(\verb{character(1}) Format string}

\item{allow_nullable}{(\code{logical(1)}) Set to \code{TRUE} if a format can be NULL
Defaults to \code{FALSE}.}
}
\value{
(\code{logical(1)}) \code{TRUE} if the format is in the supported list, else \code{FALSE}
}
\description{
Check if a given format is supported.
}
\keyword{internal}
