% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_outlier}
\alias{is_outlier}
\title{Check if a value is an outlier}
\usage{
is_outlier(x)
}
\arguments{
\item{x}{Vector of numeric values from which the outliers are to be detected.}
}
\value{
A logical vector indicating whether each value is an outlier.
}
\description{
Check if a value is an outlier
}
\keyword{internal}
