% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curves.R
\name{logistic_regresion_model}
\alias{logistic_regresion_model}
\title{Create model using nplr logistic regresion model}
\usage{
logistic_regresion_model(
  standard_curve_df,
  model_name = "MFI model",
  npars = 5,
  verbose = TRUE
)
}
\arguments{
\item{standard_curve_df}{Data frame with the standard curve data. Should contain columns: \code{dilution} and \code{MFI}}

\item{model_name}{Name of the model}

\item{npars}{Number of parameters to fit the model, by default 5 - the maximum value. \code{npars} also accepts a number of parameters to fit the model - an integer between 2 and 5, accepts also value \code{"all"}, which chooses best models from those with different number of parameters.}

\item{verbose}{If \code{TRUE} prints messages, \code{TRUE} by default}
}
\description{
This function uses the \code{nplr} package to fit the model. The model is fitted using the formula:

\deqn{y = B + \frac{T - B}{(1 + 10^{b \cdot (x_{mid} - x)})^s},}{y = B + (T - B) / (1 + 10^(b * (x_mid - x)))^s,}

where:
\itemize{
\item \eqn{y} is the predicted value, MFI in our case,
\item \eqn{x} is the independent variable, dilution in our case,
\item \eqn{B} is the bottom plateau - the right horizontal asymptote,
\item \eqn{T} is the top plateau - the left horizontal asymptote,
\item \eqn{b} is the slope of the curve at the inflection point,
\item \eqn{x_{mid}}{x_mid} is the x-coordinate at the inflection point,
\item \eqn{s} is the asymmetric coefficient.
}

This equation is refereed as the Richards' equation. More information about the model can be found in the \code{nplr} package documentation.

By default, \code{nplr} model transforms the x values using the \code{log10} function.
}
