% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Model}
\alias{Model}
\title{Logisic regresion model for the standard curve}
\description{
This function uses the \code{nplr} package to fit the model. The model is fitted using the formula:

\deqn{y = B + \frac{T - B}{(1 + 10^{b \cdot (x_{mid} - x)})^s},}{y = B + (T - B) / (1 + 10^(b * (x_mid - x)))^s,}

where:
\itemize{
\item \eqn{y} is the predicted value, MFI in our case,
\item \eqn{x} is the independent variable, dilution in our case,
\item \eqn{B} is the bottom plateau - the right horizontal asymptote,
\item \eqn{T} is the top plateau - the left horizontal asymptote,
\item \eqn{b} is the slope of the curve at the inflection point,
\item \eqn{x_{mid}}{x_mid} is the x-coordinate at the inflection point,
\item \eqn{s} is the asymmetric coefficient.
}

This equation is refereed as the Richards' equation. More information about the model can be found in the \code{nplr} package documentation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Model-new}{\code{Model$new()}}
\item \href{#method-Model-predict}{\code{Model$predict()}}
\item \href{#method-Model-plot_data}{\code{Model$plot_data()}}
\item \href{#method-Model-top_asymptote}{\code{Model$top_asymptote()}}
\item \href{#method-Model-bottom_asymptote}{\code{Model$bottom_asymptote()}}
\item \href{#method-Model-clone}{\code{Model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-new"></a>}}
\if{latex}{\out{\hypertarget{method-Model-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$new(
  dilutions,
  mfi,
  npars = 5,
  verbose = TRUE,
  log_dilution = TRUE,
  log_mfi = TRUE,
  scale_mfi = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dilutions}}{Diltuions used to fit the model}

\item{\code{mfi}}{MFI values used to fit the model}

\item{\code{npars}}{Number of parameters to use in the model}

\item{\code{verbose}}{If \code{TRUE} prints messages, \code{TRUE} by default}

\item{\code{log_dilution}}{If \code{TRUE} the dilutions are transformed using the \code{log10} function, \code{TRUE} by default}

\item{\code{log_mfi}}{If \code{TRUE} the MFI values are transformed using the \code{log10} function, \code{TRUE} by default}

\item{\code{scale_mfi}}{If \code{TRUE} the MFI values are scaled to the range [0, 1], \code{TRUE} by default}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Model-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$predict(mfi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mfi}}{MFI values used to fit the model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-Model-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$plot_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-top_asymptote"></a>}}
\if{latex}{\out{\hypertarget{method-Model-top_asymptote}{}}}
\subsection{Method \code{top_asymptote()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$top_asymptote()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-bottom_asymptote"></a>}}
\if{latex}{\out{\hypertarget{method-Model-bottom_asymptote}{}}}
\subsection{Method \code{bottom_asymptote()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$bottom_asymptote()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
