% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-plate.R
\name{Plate}
\alias{Plate}
\title{Plate}
\description{
A class to represent the luminex plate. It contains information about
the samples and analytes that were examined on the plate as well as
some additional metadata and batch info
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Plate-new}{\code{Plate$new()}}
\item \href{#method-Plate-print}{\code{Plate$print()}}
\item \href{#method-Plate-summary}{\code{Plate$summary()}}
\item \href{#method-Plate-get_data}{\code{Plate$get_data()}}
\item \href{#method-Plate-get_dilution}{\code{Plate$get_dilution()}}
\item \href{#method-Plate-get_dilution_values}{\code{Plate$get_dilution_values()}}
\item \href{#method-Plate-blank_adjustment}{\code{Plate$blank_adjustment()}}
\item \href{#method-Plate-layout}{\code{Plate$layout()}}
\item \href{#method-Plate-clone}{\code{Plate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-new"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-new}{}}}
\subsection{Method \code{new()}}{
Method to initialize the Plate object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$new(
  plate_name,
  sample_names,
  analyte_names,
  dilutions = NULL,
  dilution_values = NULL,
  sample_types = NULL,
  data = NULL,
  sample_locations = NULL,
  default_data_type = NULL,
  batch_info = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-print"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-print}{}}}
\subsection{Method \code{print()}}{
Function prints the basic information about the plate
such as the number of samples and analytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-summary}{}}}
\subsection{Method \code{summary()}}{
Function outputs basic information about the plate, such as
examination date, batch name, and sample types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$summary(..., include_names = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_names}}{If \code{include_names} parameter is \code{TRUE}, a
part from count of control samples, provides also their names.
By default \code{FALSE}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-get_data}{}}}
\subsection{Method \code{get_data()}}{
Function returns data for a specific analyte and sample.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$get_data(analyte, sample_type = "ALL", data_type = self$data_type_used)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyte}}{An analyte name or its id of which data we want to extract}

\item{\code{data_type}}{if \code{NULL} returns whole column of the dataframe
containing information about the sample. Default value is plate's
\code{data_type_used} usually \code{Median}.}

\item{\code{sample}}{sample name or id}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data about a sample and analyte
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-get_dilution"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-get_dilution}{}}}
\subsection{Method \code{get_dilution()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$get_dilution(sample_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-get_dilution_values"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-get_dilution_values}{}}}
\subsection{Method \code{get_dilution_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$get_dilution_values(sample_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-blank_adjustment"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-blank_adjustment}{}}}
\subsection{Method \code{blank_adjustment()}}{
Function adjusts the values of test samples by subtracting average of
BLANK samples purpose of this operation is to remove background light
In short it subtracts the values from data in all samples, except from
Blanks. It does not subtract values from \code{Count} values, even if this
step seems logical, people from the lab do not do it always.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$blank_adjustment(method = "avg", in_place = "TRUE")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{How the values of different blanks should be aggregated.
By default \code{avg}. For now it is the only available method}

\item{\code{inplace}}{Whether the method should produce new plate with adjusted
values or not, By default \code{TRUE} - operates on the current plate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-layout"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-layout}{}}}
\subsection{Method \code{layout()}}{
Function takes the data 1D array and returns the data in the form of
a 2D array, that represents the spatial arrangement of the samples
on the plate, ie. 12x8 grid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$layout(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{1D array of data represents for example counts, dilutions
or sample types}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 2D array representing data with the spatial arrangement
of the samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
