% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-plate_builder.R
\name{PlateBuilder}
\alias{PlateBuilder}
\title{PlateBuilder}
\description{
PlateBuilder

PlateBuilder
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlateBuilder-new}{\code{PlateBuilder$new()}}
\item \href{#method-PlateBuilder-set_sample_locations}{\code{PlateBuilder$set_sample_locations()}}
\item \href{#method-PlateBuilder-set_dilutions}{\code{PlateBuilder$set_dilutions()}}
\item \href{#method-PlateBuilder-extract_sample_types}{\code{PlateBuilder$extract_sample_types()}}
\item \href{#method-PlateBuilder-set_data}{\code{PlateBuilder$set_data()}}
\item \href{#method-PlateBuilder-set_default_data_type}{\code{PlateBuilder$set_default_data_type()}}
\item \href{#method-PlateBuilder-set_batch_info}{\code{PlateBuilder$set_batch_info()}}
\item \href{#method-PlateBuilder-set_layout}{\code{PlateBuilder$set_layout()}}
\item \href{#method-PlateBuilder-build}{\code{PlateBuilder$build()}}
\item \href{#method-PlateBuilder-clone}{\code{PlateBuilder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateBuilder-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlateBuilder-new}{}}}
\subsection{Method \code{new()}}{
Initialize the PlateBuilder object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateBuilder$new(plate_name, sample_names, analyte_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plate_name}}{\itemize{
\item plate name obtained from filename
}}

\item{\code{sample_names}}{\itemize{
\item vector of sample names measured during
an examination in the same order as in the data
}}

\item{\code{analyte_names}}{\itemize{
\item vector of analytes names measured during
an examination in the same order as in the data
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateBuilder-set_sample_locations"></a>}}
\if{latex}{\out{\hypertarget{method-PlateBuilder-set_sample_locations}{}}}
\subsection{Method \code{set_sample_locations()}}{
Set the sample types used during the examination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateBuilder$set_sample_locations(sample_locations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_locations}}{vector of sample locations pretty name ie. A1, B2}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateBuilder-set_dilutions"></a>}}
\if{latex}{\out{\hypertarget{method-PlateBuilder-set_dilutions}{}}}
\subsection{Method \code{set_dilutions()}}{
Set the dilutions (end extract numeric values) used during the examination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateBuilder$set_dilutions(dilutions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dilutions}}{vector of dilutions used during the examination
due to the nature of data it's a vector of strings,
the numeric vales are created from those strings}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateBuilder-extract_sample_types"></a>}}
\if{latex}{\out{\hypertarget{method-PlateBuilder-extract_sample_types}{}}}
\subsection{Method \code{extract_sample_types()}}{
Use the sample names to extract the sample types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateBuilder$extract_sample_types()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateBuilder-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-PlateBuilder-set_data}{}}}
\subsection{Method \code{set_data()}}{
Set the data used during the examination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateBuilder$set_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a named list of data frames containing information about
the samples and analytes. The list is named by the type of the data
e.g. \code{Median}, \verb{Net MFI}, etc.
The data frames contain information about the samples and analytes
The rows are different measures, whereas the columns represent
different analytes
Example of how \code{data$Median} looks like:\tabular{llll}{
   Sample \tab Analyte1 \tab Analyte2 \tab Analyte3 \cr
   Sample1 \tab 1.2 \tab 2.3 \tab 3.4 \cr
   Sample2 \tab 4.5 \tab 5.6 \tab 6.7 \cr
   ... \tab ... \tab ... \tab ... \cr
   Sample96 \tab 7.8 \tab 8.9 \tab 9.0 \cr
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateBuilder-set_default_data_type"></a>}}
\if{latex}{\out{\hypertarget{method-PlateBuilder-set_default_data_type}{}}}
\subsection{Method \code{set_default_data_type()}}{
Set the data type used for calculations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateBuilder$set_default_data_type(data_type = "Median")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_type}}{a character value representing the type of data
that is currently used for calculations. By default, it is set to Median}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateBuilder-set_batch_info"></a>}}
\if{latex}{\out{\hypertarget{method-PlateBuilder-set_batch_info}{}}}
\subsection{Method \code{set_batch_info()}}{
Set the batch info for the plate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateBuilder$set_batch_info(batch_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{batch_info}}{a raw list containing metadata about
the plate read from the Luminex file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateBuilder-set_layout"></a>}}
\if{latex}{\out{\hypertarget{method-PlateBuilder-set_layout}{}}}
\subsection{Method \code{set_layout()}}{
Set the layout matrix for the plate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateBuilder$set_layout(layout_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_matrix}}{a matrix containing information about the}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateBuilder-build"></a>}}
\if{latex}{\out{\hypertarget{method-PlateBuilder-build}{}}}
\subsection{Method \code{build()}}{
Create a Plate object from the PlateBuilder object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateBuilder$build(validate = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlateBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
