% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plate.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot counts in a 96-well plate}
\usage{
plot_counts(counts, antibody_name = NULL, plot_counts = FALSE)
}
\arguments{
\item{counts}{A vector with 96 counts}

\item{antibody_name}{The name of the antibody}

\item{plot_counts}{Logical indicating if the counts should be plotted}
}
\value{
A ggplot object
}
\description{
It is a function to plot counts in a 96-well plate using a color scale.
The function uses a color scale to map the counts to colors and plot them
in a 96-well plate.
The function uses the plot_plate function to plot the counts.
}
