% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plate.R
\name{plot_layout}
\alias{plot_layout}
\title{Plot layout of a 96-well plate}
\usage{
plot_layout(plate, plot_legend = TRUE)
}
\arguments{
\item{plate}{The plate object with the layout information}

\item{plot_legend}{Logical indicating if the legend should be plotted}
}
\value{
A ggplot object
}
\description{
It is a function to plot the layout of a 96-well plate using a color scale.
The function uses the plot_plate function to plot the layout.
}
\examples{
plate_filepath <- system.file("extdata", "CovidOISExPONTENT_CO.csv", package = "PvSTATEM", mustWork = TRUE)
layout_filepath <- system.file("extdata", "CovidOISExPONTENT_CO_layout.xlsx", package = "PvSTATEM", mustWork = TRUE)
plate <- read_luminex_data(plate_filepath, layout_filepath)
plot_layout(plate = plate, plot_legend = T)

}
