% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-mfi.R
\name{plot_mfi_for_analyte}
\alias{plot_mfi_for_analyte}
\title{Plot MFI value distribution for a given analyte}
\usage{
plot_mfi_for_analyte(
  plate,
  analyte_name,
  data_type = "Median",
  plot_type = "boxplot"
)
}
\arguments{
\item{plate}{A plate object}

\item{analyte_name}{The analyte to plot}

\item{data_type}{The type of data to plot. Default is "Median"}

\item{plot_type}{The type of plot to generate. Default is "boxplot".
Available options are "boxplot" and "violin".}
}
\value{
A ggplot object
}
\description{
Plot MFI value distribution for a given analyte
}
