% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curves.R
\name{plot_standard_curve_analyte_with_model}
\alias{plot_standard_curve_analyte_with_model}
\title{Plot standard curve of a certain analyte with fitted model}
\usage{
plot_standard_curve_analyte_with_model(
  plate,
  analyte_name,
  model,
  data_type = "Median",
  decreasing_dilution_order = TRUE,
  log_scale = c("all"),
  plot_asymptote = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{plate}{Plate object}

\item{analyte_name}{Name of the analyte for which we want to plot the standard curve - the same for which the model was fitted}

\item{model}{nplr object with the model}

\item{data_type}{Data type of the value we want to plot - the same datatype as in the plate file. By default equals to \code{Median}}

\item{decreasing_dilution_order}{If \code{TRUE} the dilutions are plotted in decreasing order, \code{TRUE} by default.}

\item{log_scale}{Which elements on the plot should be displayed in log scale. By default \code{"all"}. If \code{NULL} or \code{c()} no log scale is used, if \code{"all"} or \code{c("dilutions", "MFI")} all elements are displayed in log scale.}

\item{plot_asymptote}{If \code{TRUE} the asymptotes are plotted, \code{TRUE} by default}

\item{verbose}{If \code{TRUE} prints messages, \code{TRUE} by default}
}
\value{
a ggplot object with the plot
}
\description{
Function plots the values of standard curve samples and the fitted model.
}
