% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curves.R
\name{plot_standard_curve_antibody}
\alias{plot_standard_curve_antibody}
\title{PLot standard curves of plate or list of plates}
\usage{
plot_standard_curve_antibody(
  plates,
  antibody_name,
  data_type = "Net MFI",
  file_path = NULL
)
}
\arguments{
\item{plates}{A single plate object or a list of plates. Plates should contain the same number of standard curve samples and same diltuions}

\item{antibody_name}{Name of the antibody of which standard curve we want to plot.}

\item{data_type}{Data type of the value we want to plot - the same datatype as in the plate file. By default equals to \verb{Net MFI}}

\item{file_path}{where to save the output plot. If \code{NULL} the plot is displayed, \code{NULL} by default}
}
\description{
PLot standard curves of plate or list of plates
}
