% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-standard_curve.R
\name{plot_standard_curve_stacked}
\alias{plot_standard_curve_stacked}
\title{Standard curve stacked plot for levey-jennings report}
\usage{
plot_standard_curve_stacked(plate)
}
\arguments{
\item{plate}{Plate object}
}
\value{
ggplot object with the plot
}
\description{
Function generates a plot of stacked on top of each other standard curves for a given plate.
The plot is created with the levey-jennings report in mind. But it can be run by itself.
}
