% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-standard_curve.R
\name{plot_standard_curve_thumbnail}
\alias{plot_standard_curve_thumbnail}
\title{Standard curve thumbnail for report}
\usage{
plot_standard_curve_thumbnail(plate, analyte_name, data_type = "Median")
}
\arguments{
\item{plate}{Plate object}

\item{analyte_name}{Name of the analyte of which standard curve we want to plot.}

\item{data_type}{Data type of the value we want to plot - the same
types as in the plate file. By default equals to \code{median}}
}
\value{
ggplot object with the plot
}
\description{
Function generates a thumbnail of the standard curve for a given analyte.
The thumbnail is used in the plate report. It doesn't have any additional
parameters, because it is used only internally.
}
\keyword{internal}
