% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict.Model}
\alias{predict.Model}
\title{Predict the dilutions from the MFI values}
\usage{
\method{predict}{Model}(object, mfi)
}
\arguments{
\item{object}{Model object}

\item{mfi}{MFI values for which we want to predict the dilutions.
Should be in the same scale as the MFI values used to fit the model}
}
\description{
Predict the dilutions from the MFI values
}
