% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-dir.R
\name{process_dir}
\alias{process_dir}
\title{Process a dir of files to generate normalized data and reports}
\usage{
process_dir(
  input_dir,
  output_dir = NULL,
  recurse = FALSE,
  flatten_output = FALSE,
  layout_filepath = NULL,
  format = NULL,
  normalization_types = c("RAU", "nMFI"),
  generate_reports = FALSE,
  verbose = TRUE,
  dry_run = FALSE,
  return_plates = FALSE,
  ...
)
}
\description{
Process a dir of files to generate normalized data and reports
}
