% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data stored in the csv file - output of examination}
\usage{
read_data(
  file_path,
  layout_file_path = NULL,
  check_plate = TRUE,
  verbose = TRUE,
  colorize = !isTRUE(getOption("knitr.in.progress")),
  ...
)
}
\arguments{
\item{file_path}{path to the Luminex csv file}

\item{layout_file_path}{path to the layout file}

\item{check_plate}{if TRUE veryfies the plate - checks the consistency etc.}

\item{verbose}{if TRUE, print out the progress of the function}

\item{colorize}{if TRUE, colorize the output}
}
\description{
file structure based on the https://genome.med.harvard.edu/documents/luminex/IS2.3_SW_110_eng_us.pdf page 97
and Kenya screening csv files
}
\examples{
plate_file <- system.file("extdata", "random.csv", package = "PvSTATEM")
plate <- read_data(plate_file)

}
