% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-intelliflex.R
\name{read_intelliflex_format}
\alias{read_intelliflex_format}
\title{Read the Intelliflex format data}
\usage{
read_intelliflex_format(path)
}
\arguments{
\item{path}{Path to the INTELLIFLEX file}
}
\description{
Read the Intelliflex format data
}
