% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-layout.R
\name{read_layout_data}
\alias{read_layout_data}
\title{Read layout data from a file}
\usage{
read_layout_data(layout_file_path, ...)
}
\arguments{
\item{layout_file_path}{Path to the layout file}

\item{...}{Additional arguments to pass to the underlying read function}
}
\value{
A matrix with the layout data.
The row names are supposed to be letters A,B,C, etc.
The column names are supposed to be numbers 1,2,3, etc.
}
\description{
Read layout data from a file
}
