% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-xponent.R
\name{read_xponent_format}
\alias{read_xponent_format}
\title{Read the xPONENT format data}
\usage{
read_xponent_format(
  path,
  exact_parse = FALSE,
  encoding = "utf-8",
  separator = ","
)
}
\arguments{
\item{path}{Path to the xPONENT file}

\item{exact_parse}{Whether to parse the file exactly or not
Exact parsing means that the batch, calibration and assay metadata will be parsed as well}

\item{encoding}{Encoding of the file}

\item{separator}{Separator for the CSV values}
}
\description{
Read the xPONENT format data
}
