% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_empty_lists}
\alias{remove_empty_lists}
\title{Remove Empty Lists from a List}
\usage{
remove_empty_lists(lst)
}
\arguments{
\item{lst}{A list to be processed.}
}
\value{
A list with empty lists removed.
}
\description{
This internal function filters out elements from a list that are empty lists.
}
\keyword{internal}
