% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Sample}
\alias{Sample}
\title{Sample}
\description{
A class to represent the sample. It contains all the necessary
information about the sample
}
\examples{
sample <- Sample$new(id = 1, sample_name = "test",
sample_type = SampleType$new("TEST"),
sample_location = SampleLocation$new(1, 1))

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Sample id, when reading from a plate it is a id extracted from
the sample position on the plate}

\item{\code{sample_name}}{Sample name}

\item{\code{sample_type}}{A sample type - is the sample used for testing,
is it blank, positive or negative control}

\item{\code{sample_location}}{Position on the Plate}

\item{\code{warnings}}{list of warnings warnings are read from the file or
generated during the reading process}

\item{\code{errors}}{lit of errors; errors are read from the file or
generated during the reading process}

\item{\code{data}}{The dataframe containing informations about the analytes within the sample.
The dataframe rows are different measures, such as \code{Median}, \verb{Net MFI}, etc.
whereas the columns represent different analytes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sample-new}{\code{Sample$new()}}
\item \href{#method-Sample-print}{\code{Sample$print()}}
\item \href{#method-Sample-join}{\code{Sample$join()}}
\item \href{#method-Sample-clone}{\code{Sample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-new"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-new}{}}}
\subsection{Method \code{new()}}{
Initializes the sample object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$new(
  id,
  sample_name,
  sample_type = NULL,
  sample_location = NULL,
  warnings = list(),
  errors = list(),
  data = data.frame()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{numeric(1)})\cr
The id of the sample - usually just the sample row number,
can be derived from the sample location}

\item{\code{sample_name}}{(\code{character(1)})\cr
The name of the sample}

\item{\code{sample_type}}{(\code{SampleType})\cr
The type of the sample

The location of the sample on the plate}

\item{\code{warnings}}{(\code{list})\cr
List containing all the warnings connected to the sample -
generated by the validity checks or read from file}

\item{\code{errors}}{(\code{list})\cr
List containing all the critical errors connected to the sample -
generated by the validity checks or read from file}

\item{\code{data}}{(\code{data.frame})\cr
The dataframe containing the informations about analytes within the sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-print"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-print}{}}}
\subsection{Method \code{print()}}{
Prints the sample information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-join"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-join}{}}}
\subsection{Method \code{join()}}{
Joins the data of two samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$join(new_sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_sample}}{(\code{Sample})\cr
A new sample to be merged with the current sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated sample object -
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
