% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SampleLocation}
\alias{SampleLocation}
\title{SampleLocation Object}
\description{
SampleLocation defines a location of a sample on the Luminex plate
The location is defined by the row and column number
and explicitly states the location of the sample
}
\examples{

sample_location <- SampleLocation$new(col = 1, row = 4)

sample_location$location_name

sample_location <- SampleLocation$parse_sample_location("65(1,F5)")
sample_location$location_name

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{col}}{column number of sample on the Luminex plate}

\item{\code{row}}{row number of sample on the Luminex plate}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{row_letter}}{row where sample is located as a letter.
For instance \code{1} -> \code{A}}

\item{\code{location_name}}{Lcation of the sample formatted
in readable way -  \verb{(row_letter, column)}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SampleLocation-new}{\code{SampleLocation$new()}}
\item \href{#method-SampleLocation-print}{\code{SampleLocation$print()}}
\item \href{#method-SampleLocation-join}{\code{SampleLocation$join()}}
\item \href{#method-SampleLocation-clone}{\code{SampleLocation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleLocation-new"></a>}}
\if{latex}{\out{\hypertarget{method-SampleLocation-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the SampleLocation class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleLocation$new(row = NA, col = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{row}}{(\code{numeric(1)})\cr
A numeric value that represents the row number of the sample location}

\item{\code{col}}{(\code{numeric(1)})\cr
A numeric value that represents the column number of the sample location}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleLocation-print"></a>}}
\if{latex}{\out{\hypertarget{method-SampleLocation-print}{}}}
\subsection{Method \code{print()}}{
Prints the information about the sample location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleLocation$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleLocation-join"></a>}}
\if{latex}{\out{\hypertarget{method-SampleLocation-join}{}}}
\subsection{Method \code{join()}}{
Joins the data of two sample locations
Performs checks to ensure that the sample locations can be joined
Essentially checks if the two sample locations are identical or
at least one of them is empty
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleLocation$join(new_location)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_location}}{(\code{SampleLocation})\cr
The sample location object to be joined with the current sample location
object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated sample location object - an invisible object,
the outputs may be stacked together
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleLocation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SampleLocation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleLocation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
