% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SampleType}
\alias{SampleType}
\title{SampleType Object}
\description{
SampleType defines the type of the sample and the dilution factor
}
\examples{
sample_type <- SampleType$new("POSITIVE CONTROL", 1)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sample_type}}{A type of the sample.
The possible sample types are:
\enumerate{
\item (\code{"BLANK"})\cr - background sample used for testing
\item (\code{"STANDARD CURVE"})\cr - a sample used to create a standard curve
\item (\code{"POSITIVE CONTROL"})\cr - a positive control sample used for testing
\item (\code{"NEGATIVE CONTROL"})\cr - a negative control sample used for testing
\item (\code{"TEST"})\cr - the actual sample used for testing
}}

\item{\code{dilution_factor}}{A numeric value that represents
the dilution factor of the sample. Used only in the case of control samples}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{character_dilution_factor}}{the dilution factor in the format \code{1/x}
\itemize{
\item useful for printing
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SampleType-new}{\code{SampleType$new()}}
\item \href{#method-SampleType-print}{\code{SampleType$print()}}
\item \href{#method-SampleType-join}{\code{SampleType$join()}}
\item \href{#method-SampleType-clone}{\code{SampleType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleType-new"></a>}}
\if{latex}{\out{\hypertarget{method-SampleType-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the SampleType class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleType$new(sample_type, dilution_factor = NA, validate_dilution = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_type}}{(\code{character(1)})\cr
A character value that represents the type of the sample
Possible values are: \code{"BLANK"}, \code{"POSITIVE CONTROL"}, \code{"STANDARD CURVE"}
\code{"NEGATIVE CONTROL"}, \code{"TEST"}.}

\item{\code{dilution_factor}}{(\code{numeric(1)})\cr
A numeric value that represents the dilution factor of the sample.
Used only in the case of control samples}

\item{\code{validate_dilution}}{Whether to validate the object for the dilution factor.
Used in the reading scripts to dynamically assign values. Default to \code{TRUE}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleType-print"></a>}}
\if{latex}{\out{\hypertarget{method-SampleType-print}{}}}
\subsection{Method \code{print()}}{
Prints the sample type and the dilution factor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleType$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleType-join"></a>}}
\if{latex}{\out{\hypertarget{method-SampleType-join}{}}}
\subsection{Method \code{join()}}{
Joins the data of two sample types
Performs checks to ensure that the sample types can be joined
Essentially checks if the two sample types are identical or
at least one of them is empty
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleType$join(new_sample_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_sample_type}}{(\code{SampleType})\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the updated sample type object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SampleType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
