% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_columns}
\alias{select_columns}
\title{Select Columns from a DataFrame}
\usage{
select_columns(df, columns, replace_value = NA)
}
\arguments{
\item{df}{A dataframe from which columns are to be selected.}

\item{columns}{A vector of column names to select.}

\item{replace_value}{Value to use for columns that do not exist in the dataframe. Default is NA.}
}
\value{
A dataframe containing the specified columns, with missing columns filled with the replacement value.
}
\description{
Selects specified columns from a dataframe. If a column
does not exist in the dataframe, it will be added with a specified replacement value.
}
\keyword{internal}
