% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort_list_by}
\alias{sort_list_by}
\title{Sort a flat list by value}
\usage{
sort_list_by(list_obj, decreasing = FALSE, value_f = function(elem) elem)
}
\arguments{
\item{list_obj}{A list to sort}

\item{decreasing}{Should the sorting by decreasing or increasing}

\item{value_f}{Function that expects a element of the list
and returns a value to sort the list by.}
}
\description{
Sort a flat list by value
}
\keyword{internal}
