% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{verbose_cat}
\alias{verbose_cat}
\title{Verbose Cat}
\usage{
verbose_cat(..., verbose = TRUE)
}
\arguments{
\item{...}{The input to be printed.}

\item{verbose}{A logical value indicating whether the input should be printed.}
}
\description{
This function prints the input to the console if the \code{verbose} argument is \code{TRUE}.
}
\keyword{internal}
