% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{verify_character_join}
\alias{verify_character_join}
\title{Verify Character Join}
\usage{
verify_character_join(x, y)
}
\arguments{
\item{x}{A character value to be compared.}

\item{y}{A character value to be compared.}
}
\value{
\code{TRUE} if the values are equal or if either value is \code{NULL}. Returns \code{FALSE} otherwise.
}
\description{
Checks if two character values are equal. If either value is \code{NULL}, returns \code{TRUE}.
}
\keyword{internal}
