% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{verify_numeric_join}
\alias{verify_numeric_join}
\title{Verify Numeric Join}
\usage{
verify_numeric_join(x, y)
}
\arguments{
\item{x}{A numeric value to be compared.}

\item{y}{A numeric value to be compared.}
}
\value{
\code{TRUE} if the values are equal or if either value is \code{NA}. Returns \code{FALSE} otherwise.
}
\description{
Checks if two numeric values are equal. If either value is \code{NA}, returns \code{TRUE}.
}
\keyword{internal}
