% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-plate_builder.R
\name{extract_dilutions_from_layout}
\alias{extract_dilutions_from_layout}
\title{Extract dilutions from the layout representation}
\usage{
extract_dilutions_from_layout(dilutions)
}
\arguments{
\item{dilutions}{vector of dilutions used during the examination
due to the nature of data it's a vector of strings,
the numeric vales are created from those strings}
}
\description{
Extract dilution factor represented as string from vector of characters.
The matches has to be exact and the dilution factor has to be in the form of \verb{1/\\d+}
}
\keyword{internal}
