% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-plate_builder.R
\name{extract_sample_names_from_layout}
\alias{extract_sample_names_from_layout}
\title{Extract sample names from layout}
\usage{
extract_sample_names_from_layout(layout_names, locations)
}
\arguments{
\item{layout_names}{a vector of sample names from the layout file}

\item{locations}{a vector of locations in the form of A1, B2, etc.}
}
\description{
Function extracts sample names from the layout file based on the provided locations.
Function assumes that the plate is 96-well and extracts
the sample names according to the provided location strings.
}
\keyword{internal}
