% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_levey_jennings_report}
\alias{generate_levey_jennings_report}
\title{Generate a report with Levey-Jennings plots.}
\usage{
generate_levey_jennings_report(
  list_of_plates,
  filename = NULL,
  output_dir = "reports"
)
}
\arguments{
\item{list_of_plates}{A list of plate objects.}

\item{filename}{(\code{character(1)}) The name of the output file. If not
provided, the filename will be created based on the plate name
with the suffix '_report.html'.}

\item{output_dir}{(\code{character(1)}) The directory where the report
should be saved. Default is 'reports'.}
}
\value{
A report.
}
\description{
This function generates a report with Levey-Jennings plots.
The report is generated using the \code{levey_jennings_report_template.Rmd} template.
}
\keyword{internal}
