% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-plate.R
\name{is_valid_sample_type}
\alias{is_valid_sample_type}
\title{Check validity of given sample type}
\usage{
is_valid_sample_type(sample_type)
}
\arguments{
\item{sample_type}{A string representing the sample type.}
}
\value{
\code{TRUE} if the sample type is valid, \code{FALSE} otherwise.
}
\description{
Check if the sample type is valid. The sample type is valid if it is one of the
elements of the \code{VALID_SAMPLE_TYPES} vector. The valid sample types are:

\code{c(ALL, BLANK, TEST, NEGATIVE CONTROL, STANDARD CURVE, POSITIVE CONTROL)}.
}
