% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-plate.R
\name{Plate}
\alias{Plate}
\title{Plate}
\description{
A class to represent the luminex plate. It contains information about
the samples and analytes that were examined on the plate as well as
some additional metadata and batch info
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{plate_name}}{\itemize{
\item plate name obtained from filename
}}

\item{\code{analytes_names}}{vector of analytes names measured during
an examination in the same order as in the data}

\item{\code{sample_names}}{vector of sample names measured during
an examination in the same order as in the data}

\item{\code{sample_locations}}{vector of sample locations pretty name ie. A1, B2}

\item{\code{dilutions}}{vector of dilutions used during the examination
due to the nature of data it's a vector of strings, logic to transform
it to numeric has to be implemented in the future}

\item{\code{sample_types}}{vector of sample types used during the examination
those values are derived from the sample names in the Luminex file
using regular expressions
possibly in the future it will be possible to specify the sample types
manually either by passing a vector or by select it in the GUI}

\item{\code{data}}{a named list of data frames containing information about
the samples and analytes. The list is named by the type of the data
e.g. \code{Median}, \verb{Net MFI}, etc.
The data frames contain information about the samples and analytes
The rows are different measures, whereas the columns represent
different analytes
Example of how \code{data$Median} looks like:\tabular{llll}{
   Sample \tab Analyte1 \tab Analyte2 \tab Analyte3 \cr
   Sample1 \tab 1.2 \tab 2.3 \tab 3.4 \cr
   Sample2 \tab 4.5 \tab 5.6 \tab 6.7 \cr
   ... \tab ... \tab ... \tab ... \cr
   Sample96 \tab 7.8 \tab 8.9 \tab 9.0 \cr
}}

\item{\code{data_type_used}}{a character value representing the type of data
that is currently used for calculations. By default, it is set to Median}

\item{\code{batch_info}}{a raw list containing metadata about
the plate read from the Luminex file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Plate-print}{\code{Plate$print()}}
\item \href{#method-Plate-summary}{\code{Plate$summary()}}
\item \href{#method-Plate-get}{\code{Plate$get()}}
\item \href{#method-Plate-blank_adjustment}{\code{Plate$blank_adjustment()}}
\item \href{#method-Plate-layout}{\code{Plate$layout()}}
\item \href{#method-Plate-clone}{\code{Plate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-print"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-print}{}}}
\subsection{Method \code{print()}}{
Function prints the basic information about the plate
such as the number of samples and analytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-summary}{}}}
\subsection{Method \code{summary()}}{
Function outputs basic information about the plate, such as
examination date, batch name, and sample types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$summary(..., include_names = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_names}}{If \code{include_names} parameter is \code{TRUE}, a
part from count of control samples, provides also their names.
By default \code{FALSE}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-get"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-get}{}}}
\subsection{Method \code{get()}}{
Function returns data for a specific analyte and sample.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$get(analyte, sample_type, data_type = self$data_type_used)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyte}}{An analyte name or its id of which data we want to extract}

\item{\code{data_type}}{if \code{NULL} returns whole column of the dataframe
containing information about the sample. Default value is plate's
\code{data_type_used} usually \code{Median}.}

\item{\code{sample}}{sample name or id}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data about a sample and analyte
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-blank_adjustment"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-blank_adjustment}{}}}
\subsection{Method \code{blank_adjustment()}}{
Function adjusts the values of test samples by subtracting average of
BLANK samples purpose of this operation is to remove background light
In short it subtracts the values from data in all samples, except from
Blanks. It does not subtract values from \code{Count} values, even if this
step seems logical, people from the lab do not do it always.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$blank_adjustment(method = "avg", in_place = "TRUE")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{How the values of different blanks should be aggregated.
By default \code{avg}. For now it is the only available method}

\item{\code{inplace}}{Whether the method should produce new plate with adjusted
values or not, By default \code{TRUE} - operates on the current plate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-layout"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-layout}{}}}
\subsection{Method \code{layout()}}{
Function takes the data 1D array and returns the data in the form of
a 2D array, that represents the spatial arrangement of the samples
on the plate, ie. 12x8 grid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$layout(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{1D array of data represents for example counts, dilutions
or sample types}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 2D array representing data with the spatial arrangement
of the samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Plate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
