% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plate.R
\name{plot_layout}
\alias{plot_layout}
\title{Plot layout of a 96-well plate}
\usage{
plot_layout(sample_types, plate_name = NULL)
}
\arguments{
\item{sample_types}{A vector with 96 sample types}

\item{plate_name}{The name of the plate}
}
\value{
A ggplot object
}
\description{
It is a function to plot the layout of a 96-well plate using a color scale.
The function uses the plot_plate function to plot the layout.
}
