% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plate.R
\name{plot_plate}
\alias{plot_plate}
\title{Plot a 96-well plate with colored wells}
\usage{
plot_plate(colors, plot_numbers = FALSE, numbers = NULL, plot_title = "Plate")
}
\arguments{
\item{colors}{A vector with 96 colors}

\item{plot_numbers}{Logical indicating if the well numbers should be plotted}

\item{numbers}{A vector with 96 numbers}

\item{plot_title}{The title of the plot (default is "Plate")}
}
\value{
A ggplot object
}
\description{
It is a generic function to plot a 96-well plate with colored wells
used by other functions in this package, mainly to plot layout and counts.
The function uses a background image of a 96-well plate and
plots the colors in the wells using ggplot2.
}
