% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curves.R
\name{plot_standard_curve_antibody}
\alias{plot_standard_curve_antibody}
\title{Plot standard curves of plate or list of plates}
\usage{
plot_standard_curve_antibody(
  plates,
  antibody_name,
  data_type = "Median",
  decreasing_dilution_order = TRUE,
  log_scale = c("all"),
  plot_line = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{plates}{A single plate object or a list of plates. Plates should contain the same number of standard curve samples and same diltuions}

\item{antibody_name}{Name of the antibody of which standard curve we want to plot.}

\item{data_type}{Data type of the value we want to plot - the same datatype as in the plate file. By default equals to \verb{Net MFI}}

\item{decreasing_dilution_order}{If \code{TRUE} the dilutions are plotted in decreasing order, \code{TRUE} by default}

\item{log_scale}{Which elements on the plot should be displayed in log scale. By default \code{"dilutions"}. If \code{NULL} or \code{c()} no log scale is used, if \code{"all"} or \code{c("dilutions", "MFI")} all elements are displayed in log scale.}

\item{plot_line}{If \code{TRUE} a line is plotted, \code{TRUE} by default}

\item{verbose}{If \code{TRUE} prints messages, \code{TRUE} by default}
}
\value{
ggplot object with the plot
}
\description{
Plot standard curves of plate or list of plates
}
\examples{

plate_filepath <- system.file("extdata", "CovidOISExPONTENT_CO.csv", package = "PvSTATEM", mustWork = TRUE) # get the filepath of the csv dataset

plate <- read_data(plate_filepath)

plot_standard_curve_antibody(plate, antibody_name = "RBD_wuhan_IPP")

plot_standard_curve_antibody(plate, antibody_name = "RBD_wuhan_IPP", log_scale = c("dilutions")) # `log_scale` option allows to apply log scale on selected axis

plot_standard_curve_antibody(plate, antibody_name = "RBD_wuhan_IPP", decreasing_dilution_order = FALSE) # reversed x - axis

}
