% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-model.R
\name{predict.Model}
\alias{predict.Model}
\title{Predict the RAU values from the MFI values}
\usage{
\method{predict}{Model}(object, mfi, ...)
}
\arguments{
\item{object}{(\code{Model()})
Object of the Model class}

\item{mfi}{(\code{numeric()})
MFI values for which we want to predict the RAU values
Should be in the same scale as the MFI values used to fit the model}

\item{...}{Additional arguments passed to the method}
}
\value{
(\code{data.frame()})
}
\description{
More details can be found here: \link[PvSTATEM]{Model}
}
