% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-intelliflex.R
\name{read_intelliflex_format}
\alias{read_intelliflex_format}
\title{Read the Intelliflex format data}
\usage{
read_intelliflex_format(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to the INTELLIFLEX file}

\item{verbose}{Print additional information. Default is \code{TRUE}}
}
\description{
Read the Intelliflex format data
}
